Update price entity to float

+++ src/Entity/Product.php
@@ -28,7 +28,7 @@ class Product
     private ?string $productType = null;
 
     #[ORM\Column]
-    private ?int $price = null;
+    private ?float $price = null;
 
     #[ORM\Column(type: Types::TEXT)]
     private ?string $description = null;
@@ -110,12 +110,12 @@ class Product
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
